Wrap

+++ extensions/handler/stats.go
@@ -1,7 +1,7 @@
 package handler
 
 type messagesStats struct {
-	sent     int
-	failures int
-	ignored  int
+	sent     int64
+	failures int64
+	ignored  int64
 }
